Week 4 Critical Thinking Assignment Option #2: Laplacian Filters for Different Kernel Windows
Image filtering involves the application of window operations that perform useful functions, such as noise removal and image enhancement. Compare the effects of Laplacian and Gaussian filters on an image for different kernel windows.

This image Download imagecontains Gaussian noise. In OpenCV, write algorithms for this image to do the following:

Apply a Gaussian, a Laplacian, and a Gaussian with Laplacian filter using a 3x3 kernel. For Gaussian, think about how to select a good value of sigma for optimal results.
Apply a Gaussian, a Laplacian, and a Gaussian with Laplacian filter using a 5x5 kernel. For gaussian, use the same value of sigma you selected in the above step.
Apply a Gaussian, a Laplacian, and a Gaussian with Laplacian filter using a 7x7 kernel. For gaussian, use the same value of sigma you selected in the above step.
Output your filter results as 3 x 3 side-by-side subplots to make comparisons easy to inspect visually. That is, your subplot should have 3 rows (1 for each kernel size) and 3 columns (1 for each filter type). Be sure to include row and column labels.