I’ve been updating the textbook’s Jupyter notebooks for compatibility with Python 3.13. I adjusted some code to avoid deprecation errors and runtime warnings. Imports and functions were refactored accordingly to allow the notebooks to run on a clean virtual environment, requiring users only to install environment dependencies as they appear in the notebooks using PyCharm’s package installer shortcuts (pip install).
After fighting with TensorFlow for far too long, I went about this chapter's contents with PyTorch and SciKitLearn. We lose some functionality in greater stability. The tools that were compatible in 2020 just aren't anymore, this is a workaround. If you get TF and PyMC3 to work please feel free to pull